AircraftHealthML-main/requirements.txt
tensorflow
djangorestframework
django-filter
pandas
numpy
torch
scikit-learn
plotly
matplotlib
seaborn
joblib
